pyproject.toml
[tool.poetry]
name = "ifs-bot"
version = "0.1.0"
description = ""
authors = ["JeffKatzy <JeffreyEricKatz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
black = "^24.8.0"
isort = "^5.13.2"
langchain = "^0.2.12"
langchain-openai = "^0.1.20"
langgraph = "^0.2.2"
langchain-community = "^0.2.11"
langgraph-cli = "^0.1.50"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
